refactor: Split ingest code into helpers.py and main.py
- Extract helper functions into separate helpers.py module
- Move S3 operations, API requests, and data processing to helpers
- Keep main orchestration logic in main.py
- Maintain all incremental load functionality
- Improve code organization and maintainability

This refactoring makes the codebase more modular and easier to maintain while preserving all existing functionality.
